requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
certifi
deepgram-sdk
groq
python-multipart